SWE_API/requirements.txt
# Core FastAPI dependencies
fastapi==0.121.3
uvicorn==0.38.0
pydantic==2.12.4
pydantic_core==2.41.5

# HTTP and networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.11.0
sniffio==1.3.1
certifi==2025.11.12
idna==3.11

# ASGI and web server
starlette==0.50.0
asgiref==3.8.1

# C/C++ code parsing
libclang==18.1.1

# Testing
pytest==9.0.1
iniconfig==2.3.0
pluggy==1.6.0
packaging==25.0

# Type checking and annotations
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0

# Utilities
click==8.3.1
colorama==0.4.6
Pygments==2.19.2

# Optional dependencies (can be removed if not needed)
# Django==5.2.1
# sqlparse==0.5.3
# psycopg2-binary==2.9.10
# tzdata==2025.2
# annotated-doc==0.0.4

# Additional useful packages
requests>=2.31.0  # For example_workflow.py
